refactor(docker-jans): introduce hooks.py to override entrypoints
