+++ docker-jans-persistence-loader/scripts/spanner_setup.py
@@ -8,7 +8,7 @@ from jans.pycloudlib.persistence.spanner import SpannerClient
 
 from settings import LOGGING_CONFIG
 from utils import prepare_template_ctx
-from ldif_mappings import get_ldif_mappings
+from hooks import get_ldif_mappings_hook
 
 FIELD_RE = re.compile(r"[^0-9a-zA-Z\s]+")
 
@@ -158,7 +158,7 @@ class SpannerBackend:
 
     def import_builtin_ldif(self, ctx):
         optional_scopes = json.loads(self.manager.config.get("optional_scopes", "[]"))
-        ldif_mappings = get_ldif_mappings("spanner", optional_scopes)
+        ldif_mappings = get_ldif_mappings_hook("spanner", optional_scopes)
 
         for _, files in ldif_mappings.items():
             for file_ in files:
